+++ drivers/syslog/vsyslog.c
@@ -132,8 +132,8 @@ int nx_vsyslog(int priority, FAR const IPTR char *fmt, FAR va_list *ap)
 #if defined(CONFIG_SYSLOG_TIMESTAMP)
   /* Pre-pend the message with the current time, if available */
 
-  ret = lib_sprintf(&stream.public, "[%5d.%06d] ",
-                    ts.tv_sec, ts.tv_nsec / 1000);
+  ret = lib_sprintf(&stream.public, "[%5jd.%06ld] ",
+                    (uintmax_t)ts.tv_sec, ts.tv_nsec / 1000);
 #else
   ret = 0;
 #endif
